Remove path attribute from Mojo::Log->new
The `path` attribute tells Mojo::Log to write to a log file, while in a
container environment, we want the logging to go to STDOUT/STDERR that
way the centralized logging can manage it.

Removing the path attribute will also remove the error when `update.pl`
executes that it can't find the `log` directory that we don't want to
exist anyway.